% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COLOC.ABF.R
\name{COLOC.ABF}
\alias{COLOC.ABF}
\title{Perform Colocalization Analysis using Approximate Bayes Factor (ABF)}
\usage{
COLOC.ABF(
  e.dat,
  o.dat,
  e.col.chr = "CHR",
  e.col.pos = "POS",
  e.col.snp = "SNP",
  e.col.beta = "beta",
  e.col.se = "se",
  e.col.eaf = "eaf",
  e.col.pval = "pval",
  e.col.nsample = "N",
  e.col.maf = NA,
  e.type = "quant",
  e.prevalence = NA,
  o.col.snp = "SNP",
  o.col.beta = "beta",
  o.col.se = "se",
  o.col.eaf = "eaf",
  o.col.pval = "pval",
  o.col.nsample = "N",
  o.col.maf = NA,
  o.type = "cc",
  o.prevalence = NA,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  verbose = TRUE
)
}
\arguments{
\item{e.dat}{Data frame or file path for the first dataset (e.g., eQTL data).}

\item{o.dat}{Data frame or file path for the second dataset (e.g., GWAS data).}

\item{e.col.chr}{Character. Column name for chromosome in \code{e.dat}. Default: "CHR".}

\item{e.col.pos}{Character. Column name for position in \code{e.dat}. Default: "POS".}

\item{e.col.snp}{Character. Column name for SNP identifier in \code{e.dat}. Default: "SNP".}

\item{e.col.beta}{Character. Column name for effect size (beta) in \code{e.dat}. Default: "beta".}

\item{e.col.se}{Character. Column name for standard error in \code{e.dat}. Default: "se".}

\item{e.col.eaf}{Character. Column name for effect allele frequency in \code{e.dat}. Default: "eaf".}

\item{e.col.pval}{Character. Column name for p-value in \code{e.dat}. Default: "pval".}

\item{e.col.nsample}{Character or Numeric. Column name or numeric value for sample size in \code{e.dat}. Default: "N".}

\item{e.col.maf}{Character. Column name for minor allele frequency in \code{e.dat}. Default: \code{NA}.}

\item{e.type}{Character. Type of trait for \code{e.dat}, either "quant" (quantitative) or "cc" (case-control). Default: "quant".}

\item{e.prevalence}{Numeric. Disease prevalence for \code{e.dat} if \code{e.type = "cc"}. Default: \code{NA}.}

\item{o.col.snp}{Character. Column name for SNP identifier in \code{o.dat}. Default: "SNP".}

\item{o.col.beta}{Character. Column name for effect size (beta) in \code{o.dat}. Default: "beta".}

\item{o.col.se}{Character. Column name for standard error in \code{o.dat}. Default: "se".}

\item{o.col.eaf}{Character. Column name for effect allele frequency in \code{o.dat}. Default: "eaf".}

\item{o.col.pval}{Character. Column name for p-value in \code{o.dat}. Default: "pval".}

\item{o.col.nsample}{Character or Numeric. Column name or numeric value for sample size in \code{o.dat}. Default: "N".}

\item{o.col.maf}{Character. Column name for minor allele frequency in \code{o.dat}. Default: \code{NA}.}

\item{o.type}{Character. Type of trait for \code{o.dat}, either "quant" (quantitative) or "cc" (case-control). Default: "cc".}

\item{o.prevalence}{Numeric. Disease prevalence for \code{o.dat} if \code{o.type = "cc"}. Default: \code{NA}.}

\item{p1}{Numeric. Prior probability for a SNP to be associated with trait 1. Default: 1e-04.}

\item{p2}{Numeric. Prior probability for a SNP to be associated with trait 2. Default: 1e-04.}

\item{p12}{Numeric. Prior probability for a SNP to be associated with both traits. Default: 1e-05.}

\item{verbose}{Logical. Whether to print detailed logging messages during execution. Default: TRUE.}
}
\value{
A list containing the following elements:
\describe{
  \item{coloc.dat}{Full colocalization results from \code{coloc::coloc.abf}.}
  \item{summary}{Summary of posterior probabilities for each hypothesis (H0 to H4).}
  \item{results}{Detailed results per SNP, including posterior probabilities.}
  \item{priors}{Prior probabilities used in the analysis.}
  \item{assoc}{Merged dataset with z-scores for both traits.}
  \item{snp.sig}{The top SNP with the highest posterior probability for colocalization (H4).}
  \item{in_fn1}{Input data for trait 1 (SNP and p-value).}
  \item{in_fn2}{Input data for trait 2 (SNP and p-value).}
}
}
\description{
This function performs colocalization analysis between two traits using the
Approximate Bayes Factor (ABF) method as implemented in the \code{coloc} package.
It evaluates the evidence for shared causal variants between two datasets (e.g., eQTL and GWAS data).
}
\details{
This function integrates data from two traits and performs colocalization analysis
using the \code{coloc::coloc.abf} function. It supports both quantitative and
case-control traits and allows customization of column names for input datasets.
Missing MAF values are imputed from effect allele frequencies if not provided.
}
\examples{
\dontrun{

# Step 0. Load packages.
library(GWASkitR)
library(locuscomparer)

# Step 1. Perform colocalization analysis.
ABF.res <- COLOC.ABF(
  e.dat = GWASkitR::pQTLdat,
  o.dat = GWASkitR::sumstats,
  e.col.chr = "CHR",
  e.col.pos = "POS",
  e.col.snp = "SNP",
  e.col.beta = "beta",
  e.col.se = "se",
  e.col.eaf = "eaf",
  e.col.pval = "pval",
  e.col.nsample = 35373,
  e.col.maf = "MAF",
  e.type = "quant",
  e.prevalence = NA,
  o.col.snp = "SNP",
  o.col.beta = "beta",
  o.col.se = "se",
  o.col.eaf = "eaf",
  o.col.pval = "pval",
  o.col.nsample = 318014,
  o.col.maf = "MAF",
  o.type = "cc",
  o.prevalence = NA
)

# Step 2. Organize the results for visualization.
coloc.dat <- coloc_result_to_plot(
  data = ABF.res,
  plink = "/your/path/to/plink/plink.exe",
  bfile = "/your/path/to/plink/plink/1kg.v3/EUR"
)

# Step 3. Visualize the colocalization analysis results.
title1 <- "eQTL"
title2 <- "GWAS"
locuscomparer::locuscompare(
  in_fn1 = coloc.dat[["in_fn1"]],
  in_fn2 = coloc.dat[["in_fn2"]],
  title1 = title1,
  title2 = title2,
  genome = "hg19",
  snp = coloc.dat[["SNP"]],
  population = coloc.dat[["pop"]],
  combine = TRUE
)

# Step 4. Sensitivity plot
coloc::sensitivity(
  obj = ABF.res[["coloc.dat"]],
  rule = "H4 > 0.8",
  plot.manhattans = TRUE
)
}

}
