% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_result_to_plot.R
\name{coloc_result_to_plot.susie}
\alias{coloc_result_to_plot.susie}
\title{Prepare SuSiE Colocalization Results for Plotting}
\usage{
\method{coloc_result_to_plot}{susie}(x, rs, verbose = TRUE)
}
\arguments{
\item{x}{List. Output from \code{COLOC.SUSIE}, containing significant SNPs, association data,
and other colocalization results.}

\item{rs}{Character. Specific SNP ID (e.g., \code{"rs123"}) to focus on for plotting. Must follow
the format \code{"rs"} followed by numbers.}

\item{verbose}{Logical. If \code{TRUE}, log detailed information about the data preparation process.
Default: \code{TRUE}.}
}
\value{
A list containing the following elements:
  \describe{
    \item{\code{in_fn1}}{Data frame for the first trait (exposure), with columns for SNP ID (\code{rsid})
      and p-value (\code{pval}).}
    \item{\code{in_fn2}}{Data frame for the second trait (outcome), with columns for SNP ID (\code{rsid})
      and p-value (\code{pval}).}
    \item{\code{ld_matrix}}{Matrix of linkage disequilibrium (LD) correlations between SNPs.}
    \item{\code{geni.dat}}{Merged association data frame for plotting, containing marker information
      and Z-scores for both traits.}
    \item{\code{SNP}}{Character. The specified SNP ID for focus in visualization.}
    \item{\code{pop}}{Population information from the input data, if available.}
  }
}
\description{
This function processes colocalization results obtained from \code{COLOC.SUSIE} to prepare data
for visualization, focusing on a specific SNP. It extracts relevant subsets of the data and
organizes them into a format suitable for plotting tools such as \code{locuscomparer}.
}
\details{
This function validates the input SNP ID format and checks if it exists among significant SNPs identified
in the colocalization analysis. It then organizes the relevant data subsets for visualization, ensuring
compatibility with downstream plotting functions like \code{locuscomparer::locuscompare}.
}
\examples{
\dontrun{
# Assuming SUSIE.res is the output from COLOC.SUSIE
coloc.dat <- coloc_result_to_plot(
  x = SUSIE.res,
  rs = "rs1260326"
)

# Use the prepared data for visualization
locuscomparer::locuscompare(
  in_fn1 = coloc.dat[["in_fn1"]],
  in_fn2 = coloc.dat[["in_fn2"]],
  title1 = "eQTL",
  title2 = "GWAS",
  genome = "hg19",
  snp = coloc.dat[["SNP"]],
  population = coloc.dat[["pop"]],
  combine = TRUE
)
}

}
\seealso{
\code{\link{COLOC.SUSIE}}
}
\author{
[Abao Xing]
}
