% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_gwas.R
\name{format_gwas}
\alias{format_gwas}
\alias{format_gwas.default}
\title{Format GWAS Summary Statistics}
\usage{
format_gwas(sumstats, ...)

\method{format_gwas}{default}(
  sumstats,
  col.chr = NULL,
  col.pos = NULL,
  col.snp = NULL,
  col.ref = NULL,
  col.alt = NULL,
  col.metrics = NULL,
  col.eaf = NULL,
  col.se = NULL,
  col.pval = NULL,
  col.ncase = NULL,
  col.nsample = NULL,
  col.chr.set = "CHR",
  col.pos.set = "POS",
  col.snp.set = "SNP",
  col.ref.set = "other_allele",
  col.alt.set = "effect_allele",
  col.metrics.set = col.metrics,
  col.eaf.set = "eaf",
  col.se.set = "se",
  col.pval.set = "pval",
  col.ncase.set = "ncase",
  col.nsample.set = "nsample",
  save.path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sumstats}{An object to be formatted. For the default method, this can be a character
string (path to a summary statistics file in VCF or tabular format) or a
\code{data.frame}/\code{data.table} object containing the data.}

\item{...}{Additional arguments passed to specific methods.}

\item{col.chr}{Character string or \code{NULL}. Name of the chromosome column
in the input data.}

\item{col.pos}{Character string or \code{NULL}. Name of the position column in
the input data.}

\item{col.snp}{Character string or \code{NULL}. Name of the SNP identifier column
in the input data.}

\item{col.ref}{Character string or \code{NULL}. Name of the reference allele
column in the input data.}

\item{col.alt}{Character string or \code{NULL}. Name of the alternative allele
column in the input data.}

\item{col.metrics}{Character string or \code{NULL}. Name of the effect size column
in the input data (e.g., beta, OR, ES, Z).}

\item{col.eaf}{Character string or \code{NULL}. Name of the effect allele frequency
column in the input data.}

\item{col.se}{Character string or \code{NULL}. Name of the standard error column
in the input data.}

\item{col.pval}{Character string or \code{NULL}. Name of the p-value column in
the input data.}

\item{col.ncase}{Character string or \code{NULL}. Name of the case sample size
column in the input data.}

\item{col.nsample}{Character string or \code{NULL}. Name of the total sample size
column in the input data.}

\item{col.chr.set}{Character string. Desired output column name for chromosome.
Default is \code{"CHR"}.}

\item{col.pos.set}{Character string. Desired output column name for position.
Default is \code{"POS"}.}

\item{col.snp.set}{Character string. Desired output column name for SNP.
Default is \code{"SNP"}.}

\item{col.ref.set}{Character string. Desired output column name for reference allele.
Default is \code{"other_allele"}.}

\item{col.alt.set}{Character string. Desired output column name for alternative allele.
Default is \code{"effect_allele"}.}

\item{col.metrics.set}{Character string. Desired output column name for effect size.
Default is the value of \code{col.metrics}.}

\item{col.eaf.set}{Character string. Desired output column name for effect allele frequency.
Default is \code{"eaf"}.}

\item{col.se.set}{Character string. Desired output column name for standard error.
Default is \code{"se"}.}

\item{col.pval.set}{Character string. Desired output column name for p-value.
Default is \code{"pval"}.}

\item{col.ncase.set}{Character string. Desired output column name for case sample size.
Default is \code{"ncase"}.}

\item{col.nsample.set}{Character string. Desired output column name for total sample size.
Default is \code{"nsample"}.}

\item{save.path}{Character string or \code{NULL}. File path to save the formatted
summary statistics as a tab-separated file. If \code{NULL}, no file is saved.
Default is \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE}, prints progress and debugging messages
using the \code{logger} package. Default is \code{TRUE}.}
}
\value{
A \code{tibble} object containing the formatted summary statistics with
  selected and renamed columns.
}
\description{
A generic function to read, validate, select, and rename columns in GWAS summary
statistics files. It supports both VCF and tabular formats, ensuring required
columns exist, applying user-specified renaming, optionally saving the processed
table, and logging each step using the \code{logger} package.
}
\details{
This is a generic function for formatting GWAS summary statistics. The default
method processes data from VCF or tabular files, or directly from a data frame.
It validates the presence of specified input columns, selects only the required
columns, and renames them to user-defined output names. VCF files are read using
an internal \code{read.vcf} function, while tabular files are read using
\code{data.table::fread}. The function logs all steps and can save the formatted
data to disk if a \code{save.path} is provided.
}
\examples{
\dontrun{
library(GWASkitR)

# Example 1: Format a tabular file
sumstats.file <- system.file("extdata", "GWAS-test.txt", package = "GWASkitR")
sumstats.dt <- format_gwas(
  sumstats = sumstats.file,
  col.chr = "chr",
  col.pos = "pos",
  col.snp = "SNP",
  col.ref = "other_allele",
  col.alt = "effect_allele",
  col.metrics = "beta",
  col.eaf = "eaf",
  col.se = "se",
  col.pval = "pval",
  col.nsample = "samplesize",
  col.chr.set = "CHR",
  col.pos.set = "POS",
  col.snp.set = "SNP",
  col.ref.set = "other_allele",
  col.alt.set = "effect_allele",
  col.metrics.set = "beta",
  col.eaf.set = "eaf",
  col.se.set = "se",
  col.pval.set = "pval",
  col.nsample.set = "nsamples",
  verbose = TRUE
)
head(sumstats.dt)

# Example 2: Format a data frame
sumstats.dt <- format_gwas(
  sumstats = GWASkitR::sumstats,
  col.chr = "CHR",
  col.pos = "POS",
  col.snp = "SNP",
  col.ref = "other_allele",
  col.alt = "effect_allele",
  col.metrics = "beta",
  col.eaf = "eaf",
  col.se = "se",
  col.pval = "pval",
  col.chr.set = "CHR",
  col.pos.set = "POS",
  col.snp.set = "SNP",
  col.ref.set = "other_allele",
  col.alt.set = "effect_allele",
  col.metrics.set = "beta",
  col.eaf.set = "eaf",
  col.se.set = "se",
  col.pval.set = "pval",
  verbose = TRUE
)
head(sumstats.dt)

# Example 3: Format a VCF file
vcf.file <- system.file("extdata", "ieu-a-2.vcf.gz", package = "GWASkitR")
sumstats.dt <- format_gwas(
  sumstats = vcf.file,
  col.chr = "CHR",
  col.pos = "POS",
  col.snp = "ID",
  col.ref = "REF",
  col.alt = "ALT",
  col.metrics = "ES",
  col.eaf = "AF",
  col.se = "SE",
  col.pval = "PVAL",
  col.nsample = "SS",
  col.chr.set = "CHR",
  col.pos.set = "POS",
  col.snp.set = "SNP",
  col.ref.set = "other_allele",
  col.alt.set = "effect_allele",
  col.metrics.set = "beta",
  col.eaf.set = "eaf",
  col.se.set = "se",
  col.pval.set = "pval",
  col.nsample.set = "nsample",
  verbose = TRUE
)
head(sumstats.dt)
}

}
