% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.valid.R
\name{is.valid}
\alias{is.valid}
\title{Check if a list of column names is valid}
\usage{
is.valid(...)
}
\arguments{
\item{...}{One or more column names as character values.}
}
\value{
Logical. Returns `TRUE` if all provided column names are not NA or empty strings;
returns `FALSE` otherwise, or if no arguments are supplied.
}
\description{
This function checks whether the input column names (passed as separate arguments) are non-empty,
do not contain NA values, and do not contain empty strings.
}
\examples{
is.valid("a", "b")           # TRUE
is.valid("a", "")            # FALSE
is.valid()                   # FALSE
is.valid(NA, "b")            # FALSE
is.valid("a", NULL)          # FALSE

}
