% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.setnames.R
\name{safe.setnames}
\alias{safe.setnames}
\title{Safely Rename Columns in a Data Table}
\usage{
safe.setnames(dt, old, new, verbose = FALSE)
}
\arguments{
\item{dt}{Data table, the input data table whose column needs to be renamed.}

\item{old}{Character, the current name of the column to be renamed.}

\item{new}{Character, the new name for the column.}

\item{verbose}{Logical, whether to log information about the renaming operation (default: FALSE).}
}
\value{
Invisible, the modified data table with the renamed column (if applicable).
}
\description{
This function safely renames a column in a data table using \code{data.table::setnames},
but only if the old and new column names are different. It provides optional verbose
logging to track renaming operations.
}
\examples{
\dontrun{
  library(data.table)
  dt <- data.table(a = 1:3)
  safe.setnames(dt, old = "a", new = "b", verbose = TRUE)
  print(dt)
}

}
